+++ modules/actor/beast-actor-sheet.js
@@ -5,7 +5,7 @@ export class HMBeastActorSheet extends HMActorSheet {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             classes: ['hackmaster', 'sheet', 'actor'],
-            width: 640,
+            width: 665,
             height: 770,
             tabs: [{ navSelector: '.sheet-tabs', contentSelector: '.sheet-body', initial: 'overview' }],
         });