+++ modules/actor/character-actor-sheet.js
@@ -7,9 +7,8 @@ export class HMCharacterActorSheet extends HMActorSheet {
         return mergeObject(super.defaultOptions, {
             classes: ['hackmaster', 'sheet', 'actor'],
             width: 835,
-            height: 945,
+            height: 941,
             tabs: [{ navSelector: '.sheet-tabs', contentSelector: '.sheet-body', initial: 'skills' }],
-            tabs: [{ navSelector: '.sheet-tabs', contentSelector: '.sheet-body', initial: 'biography' }],
         });
     }